requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
langchain
langchain-community
langchain-huggingface
langchain-ollama
sentence-transformers
transformers
torch
requests
pypdf
chromadb
streamlit
pymongo
